+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- AS1/2 Better unresolved constant handling - §§constant(xx) func instead of §§constantxx
 
 ## [11.1.0] - 2018-05-24
 ### Added